requirements.txt
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
requests==2.31.0
cloudinary==1.40.0

# Flask web app
flask==3.0.0
flask-cors==4.0.0